Safari-609.1.11.1/LayoutTests/http/tests/security/cross-frame-access-history-prototype-expected.txt

CONSOLE MESSAGE: line 5: SecurityError: Blocked a frame with origin "http://localhost:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
PASS: History object prototype not tainted.
